+++ packages/tiled-maps/src/IMap2DLayer.ts
@@ -1,0 +1,8 @@
+import {IMap2DTileRenderer} from './IMap2DTileRenderer';
+
+export interface IMap2DLayer {
+  addTileRenderer(renderer: IMap2DTileRenderer): void;
+  removeTileRenderer(renderer: IMap2DTileRenderer): void;
+
+  update(): void;
+}