+++ packages/tiled-maps/src/index.ts
@@ -1,4 +1,5 @@
 export * from './AABB2';
+export * from './IMap2DLayer';
 export * from './IMap2DTileDataProvider';
 export * from './IMap2DTileRenderer';
 export * from './Map2DLayer';
@@ -8,7 +9,7 @@ export * from './Map2DTile';
 export * from './Map2DTileCoordsUtil';
 export * from './RepeatingTilesProvider';
 export * from './TileSprites/descriptors';
-export * from './TileSprites/Map2DTileSpritesRenderer';
+export * from './TileSprites/Map2DTileSprites';
 export * from './TileSprites/TileSprites';
 export * from './TileSprites/TileSpritesGeometry';
 export * from './TileSprites/TileSpritesMaterial';