+++ packages/picimo/CHANGELOG.md
@@ -1,4 +1,15 @@
-@spearwolf/picimo
+__@spearwolf/picimo__
+
+## 0.1.2
+
+- integrate _@spearwolf/tiled-maps_ package
+  - add new components:
+    - `<RepeatingTilesProvider>`
+    - `<Map2DLayer3D>`
+    - `<Map2DTileSpritesRenderer>`
+    - `<TileSpritesGeometry>`
+    - `<TileSpritesMaterial>`
+
 
 ## 0.1.1